requirements.txt
httplib2==0.9.2
oauthlib==2.0.1
python-instagram==1.3.2
requests==2.12.4
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
tweepy==3.5.0
